+++ banyand/liaison/grpc/stream.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/apache/skywalking-banyandb/api/common"
 	"github.com/apache/skywalking-banyandb/api/data"
+	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	streamv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/stream/v1"
 	"github.com/apache/skywalking-banyandb/banyand/tsdb"
 	"github.com/apache/skywalking-banyandb/pkg/accesslog"
@@ -56,8 +57,8 @@ func (s *streamService) activeIngestionAccessLog(root string) (err error) {
 }
 
 func (s *streamService) Write(stream streamv1.StreamService_WriteServer) error {
-	reply := func(stream streamv1.StreamService_WriteServer, logger *logger.Logger) {
-		if errResp := stream.Send(&streamv1.WriteResponse{}); errResp != nil {
+	reply := func(metadata *commonv1.Metadata, status streamv1.Status, stream streamv1.StreamService_WriteServer, logger *logger.Logger) {
+		if errResp := stream.Send(&streamv1.WriteResponse{Metadata: metadata, Status: status}); errResp != nil {
 			logger.Err(errResp).Msg("failed to send response")
 		}
 	}
@@ -76,18 +77,29 @@ func (s *streamService) Write(stream streamv1.StreamService_WriteServer) error {
 		}
 		if err != nil {
 			s.sampled.Error().Stringer("written", writeEntity).Err(err).Msg("failed to receive message")
-			reply(stream, s.sampled)
+			reply(nil, streamv1.Status_STATUS_RECEIVE_ERROR, stream, s.sampled)
 			continue
 		}
 		if errTime := timestamp.CheckPb(writeEntity.GetElement().Timestamp); errTime != nil {
 			s.sampled.Error().Stringer("written", writeEntity).Err(errTime).Msg("the element time is invalid")
-			reply(stream, s.sampled)
+			reply(writeEntity.GetMetadata(), streamv1.Status_STATUS_INVALID_TIMESTAMP, stream, s.sampled)
+			continue
+		}
+		streamCache, existed := s.entityRepo.getLocator(getID(writeEntity.GetMetadata()))
+		if !existed {
+			s.sampled.Error().Err(err).Stringer("written", writeEntity).Msg("failed to get stream cache")
+			reply(writeEntity.GetMetadata(), streamv1.Status_STATUS_INVALID_METADATA, stream, s.sampled)
+			continue
+		}
+		if writeEntity.Metadata.ModRevision != streamCache.ModRevision {
+			s.sampled.Error().Stringer("written", writeEntity).Msg("the stream mod revision is invalid")
+			reply(writeEntity.GetMetadata(), streamv1.Status_STATUS_EXPIRED_REVISION, stream, s.sampled)
 			continue
 		}
 		entity, tagValues, shardID, err := s.navigate(writeEntity.GetMetadata(), writeEntity.GetElement().GetTagFamilies())
 		if err != nil {
 			s.sampled.Error().Err(err).RawJSON("written", logger.Proto(writeEntity)).Msg("failed to navigate to the write target")
-			reply(stream, s.sampled)
+			reply(writeEntity.GetMetadata(), streamv1.Status_STATUS_INTERNAL_ERROR, stream, s.sampled)
 			continue
 		}
 		if s.ingestionAccessLog != nil {
@@ -108,7 +120,7 @@ func (s *streamService) Write(stream streamv1.StreamService_WriteServer) error {
 		if errWritePub != nil {
 			s.sampled.Error().Err(errWritePub).RawJSON("written", logger.Proto(writeEntity)).Msg("failed to send a message")
 		}
-		reply(stream, s.sampled)
+		reply(writeEntity.GetMetadata(), streamv1.Status_STATUS_SUCCEED, stream, s.sampled)
 	}
 }