+++ CHANGES.md
@@ -15,6 +15,7 @@ Release Notes.
 - Implement the remote queue to spreading data to data nodes.
 - Fix parse environment variables error
 - Implement the distributed query engine.
+- Add mod revision check to write requests.
 
 ### Bugs